implemented user auth for services and added new services as per requiremenet

+++ ProductCatalogService/src/main/java/org/example/productcatalogservice/controller/ProductController.java
@@ -1,14 +1,18 @@
 package org.example.productcatalogservice.controller;
 
+import org.example.productcatalogservice.client.UserServiceClient;
 import org.example.productcatalogservice.dto.CategoryDto;
 import org.example.productcatalogservice.dto.ProductDto;
+import org.example.productcatalogservice.dto.SearchRequestDto;
+import org.example.productcatalogservice.dto.UserDto;
+import org.example.productcatalogservice.exceptions.UserNotAuthenticatedException;
+import org.example.productcatalogservice.exceptions.UserNotAuthorised;
 import org.example.productcatalogservice.model.Category;
 import org.example.productcatalogservice.model.Product;
 import org.example.productcatalogservice.service.IProductService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.ParameterResolutionDelegate;
+import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
@@ -23,22 +27,37 @@ public class ProductController {
 
     @Autowired
     private IProductService productService;
+    @Autowired
+    private UserServiceClient userServiceClient;
 
     @GetMapping("/{pid}/{uid}")
-    public ProductDto getProductBasedOnUserScope(@PathVariable Long pid,@PathVariable Long uid){
+    public ResponseEntity<ProductDto> getProductBasedOnUserScope(@PathVariable Long pid,@PathVariable Long uid){
         Product product = productService.getProductBasedOnUserScope(pid,uid);
-        return from(product);
+        if(product == null)
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        return new ResponseEntity<>(from(product),HttpStatus.OK);
     }
 
     @GetMapping
-    public List<ProductDto> getProducts() {
+    public ResponseEntity<List<ProductDto>> getProducts() {
         List<Product> products = productService.getAllProducts();
         List<ProductDto> productDtos = new ArrayList<>();
         for (Product product : products) {
             ProductDto productDto = from(product);
             productDtos.add(productDto);
         }
-        return productDtos;
+        return new ResponseEntity<>(productDtos,HttpStatus.OK);
+    }
+
+    @GetMapping("/category/{categoryId}")
+    public ResponseEntity<List<ProductDto>> getProductsByCategoryId(@PathVariable Long categoryId) {
+        List<Product> products = productService.getAllProductsByCategory(categoryId);
+        List<ProductDto> productDtos = new ArrayList<>();
+        for (Product product : products) {
+            ProductDto productDto = from(product);
+            productDtos.add(productDto);
+        }
+        return new ResponseEntity<>(productDtos,HttpStatus.OK);
     }
 
     @GetMapping("{id}")
@@ -54,7 +73,7 @@ public class ProductController {
                 return new ResponseEntity<>(null,headers, HttpStatus.INTERNAL_SERVER_ERROR);
             }
             ProductDto productDto = from(product);
-            return new ResponseEntity<ProductDto>(productDto,headers,HttpStatus.OK);
+            return new ResponseEntity<>(productDto,headers,HttpStatus.OK);
         }catch (IllegalArgumentException exception){
             return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
             //throw exception;
@@ -62,33 +81,61 @@ public class ProductController {
     }
 
     @PostMapping
-    public ProductDto createProduct(@RequestBody ProductDto product) {
+    public ResponseEntity<ProductDto> createProduct(@RequestBody ProductDto product,@RequestHeader("Authorization") String token) {
+        UserDto user = userServiceClient.validateToken(token);
+        if(user == null){
+            throw new UserNotAuthenticatedException("Token is not valid");
+        }
+        if(!userServiceClient.hasRequiredRole(user, "ADMIN")){
+            throw new UserNotAuthorised("You do not have the required role");
+        }
         Product product1 = from(product);
         Product response = productService.createProduct(product1);
-        return from(response);
+        return new ResponseEntity<>(from(response),HttpStatus.CREATED);
     }
 
     @PutMapping("{id}")
-    public ProductDto  replaceProduct(@PathVariable long id,@RequestBody ProductDto productDto) {
+    public ResponseEntity<ProductDto>  replaceProduct(@PathVariable long id,@RequestBody ProductDto productDto,@RequestHeader("Authorization") String token) {
+        UserDto user = userServiceClient.validateToken(token);
+        if(user == null){
+            throw new UserNotAuthenticatedException("Token is not valid");
+        }
+        if(!userServiceClient.hasRequiredRole(user, "ADMIN")){
+            throw new UserNotAuthorised("You do not have the required role");
+        }
         Product product = from(productDto);
         Product result = productService.replaceProduct(id,product);
-        return from(result);
+        return new ResponseEntity<>(from(result),HttpStatus.OK);
     }
 
     @DeleteMapping("{id}")
-    public ProductDto deleteProduct(@PathVariable long id) {
+    public ResponseEntity<ProductDto> deleteProduct(@PathVariable long id,@RequestHeader("Authorization") String token) {
+        UserDto user = userServiceClient.validateToken(token);
+        if(user == null){
+            throw new UserNotAuthenticatedException("Token is not valid");
+        }
+        if(!userServiceClient.hasRequiredRole(user, "ADMIN")){
+            throw new UserNotAuthorised("You do not have the required role");
+        }
         Product product = productService.deleteProduct(id);
-        return from(product);
+        return new ResponseEntity<>(from(product),HttpStatus.OK);
     }
 
     @GetMapping("fill/{count}")
-    public List<ProductDto> populateProductsFromFakeStore(@PathVariable int count){
+    public ResponseEntity<List<ProductDto>> populateProductsFromFakeStore(@PathVariable int count){
         List<Product> products = productService.populateFromFakeStore(count);
         List<ProductDto> productDtos = new ArrayList<>();
         for (Product product : products) {
             productDtos.add(from(product));
         }
-        return productDtos;
+        return new ResponseEntity<>(productDtos,HttpStatus.OK);
+    }
+
+    @PostMapping("/search")
+    public Page<Product> searchProducts(@RequestBody SearchRequestDto searchRequestDto){
+        Page<Product> productList =  productService.searchProducts(searchRequestDto.getQuery(),
+                searchRequestDto.getPageNumber(), searchRequestDto.getPageSize(),searchRequestDto.getSortParamList());
+        return productList;
     }
 
     private Product from(ProductDto productDto){
